Move out the groups server servlets to a separate file.
